CrearEscenario loads from Game class
CrearEscenario loads from Game class and now SceneName is saved in
stadiums.xml (in each stadium) instead of in each level.

The class Game has no cameraFollowsPlayer variable because, after some
tests, I was unable to change this variable value. So, I decided not to
add it and look for an alternative solution.